% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_pps_survey.R
\name{pc_pps_survey}
\alias{pc_pps_survey}
\title{Sample units with proportional to size of the primary unit}
\usage{
pc_pps_survey(data, nsize = 20)
}
\arguments{
\item{data}{A data frame object. The data frame must have a column that specifies strata,
named stratum, a column that specifies clusters, named cluster, and a column that has size
of the cluster, named cluster_pop.}

\item{nsize}{Number of units to be sampled. If a single number is specified, then equal
number of clusters will be sampled from each stratum. If a vector is fed instead, then
number of units sampled will depend on the order of the strata in the data frame. To avoid confusions, you must sort your data by stratum before}
}
\value{
Returns the the data frame object with additional rows that mark selected units, probability of select, etc
}
\description{
Sample units with proportional to size of the primary unit
}
\examples{
set.seed(1000)
stratum = rep(c(1,2),c(30,50))
cluster <- c(1:30,1:50)
cluster_pop <- sample(365:1309,80,replace=TRUE)
sample_frame <- data.frame(stratum,cluster, cluster_pop)
pps_sample1 <- pc_pps_survey(data = sample_frame,nsize = 10)
pps_sample2 <- pc_pps_survey(data = sample_frame,nsize = c(10,7))
}
