% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_sample_hh_survey.R
\name{pc_sample_hh_survey}
\alias{pc_sample_hh_survey}
\title{Sample units using systematic random sampling}
\usage{
pc_sample_hh_survey(data, nsize = 20)
}
\arguments{
\item{data}{A data frame object}

\item{nsize}{Number of units to be sampled in each cluster}
}
\value{
Returns the the data frame object with additional rows that mark selected units, probability of select, etc
}
\description{
Sample units using systematic random sampling
}
