% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_pps_survey.R
\name{pc_pps_survey}
\alias{pc_pps_survey}
\title{Sample units with proportional to size of the primary unit}
\usage{
pc_pps_survey(data, nsize = 20)
}
\arguments{
\item{data}{A data frame object}

\item{nsize}{Number of units to be sampled}
}
\value{
Returns the the data frame object with additional rows that mark selected units, probability of select, etc
}
\description{
Sample units with proportional to size of the primary unit
}
\examples{
set.seed(1000)
stratum = rep(c(1,2),c(30,50))
cluster <- c(1:30,1:50)
cluster_pop <- sample(365:1309,80,replace=TRUE)
sample_frame <- data.frame(stratum,cluster, cluster_pop)
pc_pps_survey(sample_frame,10)
sample <- pc_pps_survey(sample_frame,nsize = c(10,7))



}
